google-enforced targetSdkVersion -> 29 change
That's minimal allowed as of this moment.
The debug apk wouldn't run on emu 29+,
it doesn't run on my Xiaomi 12, although it runs on one older Samsung
The compileSdkVersion bumped up to see what's going on.
The app uses long deprecated API, we need to take care of this now.